I
I changed some stuff then got mad because i couldnt resize an image that should be very simple to update and my partner figured it out and GUESS WHAt it didnt work on mine. frikkin shocking. Whoever makes these programs needs to make changing image sizes easier because its always  a problem. Saturday all i did was break the website because of the image and here we are again!